refactor(bdk): split pg cache and lookup modules
